# Speedup 05: Retrieve unique detections in family and in `matched_filter`