--- a/eqcorrscan/core/match_filter/family.py
+++ b/eqcorrscan/core/match_filter/family.py
@@ -328,8 +328,6 @@ class Family(object):
         """
         _detections = []
         _detections = list(set(self.detections))
-        # [_detections.append(d) for d in self.detections
-        # if not _detections.count(d)]
         self.detections = _detections
         return self