--- a/eqcorrscan/core/match_filter/family.py
+++ b/eqcorrscan/core/match_filter/family.py
@@ -326,8 +326,8 @@ class Family(object):
         2
         """
         _detections = []
-        _detections = [_detections.append(d) for d in self.detections
-                       if not _detections.count(d)]
+        [_detections.append(d) for d in self.detections
+         if not _detections.count(d)]
         self.detections = _detections
         return self